First Commit
Included in this commit is the .csv file that contains all of the
features and labels. Descriptions of each feature can be found in
iris.txt. Currently, the jupyter notebook contains code to import
several fundamental libraries including numpy, pandas and matplotlib.
Pandas successfully reads in the data set and stores it in a dataframe.

.gitignore contains a single entry, .ipynb_checkpoints. This directory
contains checkpoints made by jupyter. Ordinarily, keeping checkpoints is
useful for reverting back to previous revisions. However, git performs
that task for us, so it is no longer needed. Created checkpoint files
may stil prove useful to revert local changes if need be.